refactor(sip): extract SipCallSessionSubscriptionService to keep the inbound service under 1000 lines (#158)
P1-6 (26ca0d84) grew SipCallSessionInboundService.cs from 952 to 1003 lines, tripping the ArchitectureTests
R3 max-1000-lines gate. Move the RFC 6665 in-dialog event-subscription responsibility — inbound SUBSCRIBE
acceptance/lifecycle, inbound NOTIFY acknowledgement, and the accepted/expired-subscription NOTIFYs — into a
new injected collaborator SipCallSessionSubscriptionService, mirroring the existing SipReferHandler extraction.

The inbound service now delegates NOTIFY/SUBSCRIBE to _subscriptions and owns/disposes it like _refer; the
moved method bodies are byte-identical. No behaviour change. Inbound service drops to 806 lines.

Gates: ArchitectureTests 13/13, Core IntegrationTests 2046/2046 net10.
